+++ activeobjects-core/test/net/java/ao/db/sqlserver/create-table.sql
@@ -4,11 +4,11 @@ CREATE TABLE dbo.person (
     lastName NTEXT,
     age INTEGER,
     url VARCHAR(767) CONSTRAINT U_person_url UNIQUE NOT NULL,
-    height DOUBLE CONSTRAINT df_person_height DEFAULT 62.3,
+    height FLOAT CONSTRAINT df_person_height DEFAULT 62.3,
     companyID BIGINT,
     cool BIT CONSTRAINT df_person_cool DEFAULT 1,
     modified DATETIME,
-    weight DOUBLE,
+    weight FLOAT,
     CONSTRAINT fk_person_companyid FOREIGN KEY (companyID) REFERENCES dbo.company(id),
     PRIMARY KEY(id)
 )
